add messaging and finish item routes

+++ server/routes/items.js
@@ -4,6 +4,7 @@ const express = require("express"),
   db = require("../function"),
   aws = require("aws-sdk"),
   { v4: uuidv4 } = require("uuid"),
+  _ = require("lodash"),
   bodyParser = require("body-parser");
 const router = express.Router();
 
@@ -17,21 +18,15 @@ router.post(
     body("price").exists().isFloat(),
     body("location").exists().isString(),
     body("status").exists().isString(),
-    body("description").exists().isString()
+    body("description").exists().isString(),
   ],
   //customValidation.isLoggedIn,
   customValidation.validate,
   (req, res) => {
-    db.createItem("items", {
-      id: { S: uuidv4() },
-      title: { S: req.body.title },
-      seller_id: { S: req.body.seller_id },
-      price: { N: req.body.price },
-      location: { S: req.body.location },
-      status: { S: req.body.status },
-      description: {S: req.body.description}
-    });
-    res.status(201).send();
+    let new_id = uuidv4();
+    let new_item = _.extend(req.body, { id: new_id });
+    db.createItem("items", aws.DynamoDB.Converter.marshall(new_item));
+    res.status(201).json({ id: new_id });
   }
 );
 
@@ -45,8 +40,12 @@ router.get("/", async (req, res) => {
   res.status(201).json(output);
 });
 
-router.get("/:item_id", (req, res) => {
-  res.status(201).json();
+router.get("/:item_id", async (req, res) => {
+  let item = await db.getItem("items", req.params.item_id);
+  if (_.isUndefined(item.Item)) {
+    return res.status(404).json({ Error: "No item with that item_id exists" });
+  }
+  res.status(201).json(aws.DynamoDB.Converter.unmarshall(item.Item));
 });
 
 router.put(
@@ -57,20 +56,36 @@ router.put(
     body("price").exists().isFloat(),
     body("location").exists().isString(),
     body("status").exists().isString(),
-    body("description").exists().isString()
+    body("description").exists().isString(),
   ],
-  customValidation.isLoggedIn,
+  //customValidation.isLoggedIn,
   customValidation.validate,
-  (req, res) => {
-    res.status(200).send();
+  async (req, res) => {
+    let item = await db.getItem("items", req.params.item_id);
+    if (_.isUndefined(item.Item)) {
+      return res
+        .status(404)
+        .json({ Error: "No item with that item_id exists" });
+    }
+    let update = _.extend(req.body, { id: req.params.item_id });
+    item = await db.updateItem(
+      "items",
+      aws.DynamoDB.Converter.marshall(_.extend(req.body, update))
+    );
+    res.status(200).json(update);
   }
 );
 
 router.delete(
   "/:item_id",
-  customValidation.isLoggedIn,
-  customValidation.validate,
-  (req, res) => {
+  //customValidation.isLoggedIn,
+  async (req, res) => {
+    let item = await db.getItem("items", req.params.item_id);
+    if (_.isUndefined(item.Item)) {
+      return res
+        .status(404)
+        .json({ Error: "No item with this item_id exists" });
+    }
     db.deleteItem("items", req.params.item_id);
     res.status(204).send();
   }
